third_party/blink/web_tests/external/wpt/fetch/api/request/request-disturbed.any-expected.txt

This is a testharness.js-based test.
[FAIL] Input request used for creating new request became disturbed
  assert_equals: body should not change expected object "[object ReadableStream]" but got object "[object ReadableStream]"
[FAIL] Input request used for creating new request became disturbed even if body is not used
  assert_equals: body should not change expected object "[object ReadableStream]" but got object "[object ReadableStream]"
Harness: the test ran to completion.
